pyproject.toml
[tool.poetry]
name = "tig-cogs"
version = "0.1.0"
description = ""
authors = ["tigattack <10629864+tigattack@users.noreply.github.com>"]
license = "GPL-3.0-only"
readme = "README.md"
package-mode = false

[project]
requires-python = "<3.12,>=3.9"

[project.urls]
Repository = "https://github.com/tigattack/tig-cogs"
Issues = "https://github.com/tigattack/tig-cogs/issues"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
celtictuning = {git = "https://github.com/tigattack/CelticTuning"}
dvla-vehicle-enquiry-service = "^1.0.2"
dvsa-mot-history = "^1.0.2"
pillow = "^10.4.0"
python-dateutil = "^2.9.0-post0"
python3-discogs-client = "^2.7"
Red-DiscordBot = "^3.5.12"

pydantic = "^2.8.2"
[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
pre-commit = "^3.8.0"
ruff = "^0.6.1"
types-beautifulsoup4 = "^4.12.0.20240511"
types-cffi = "^1.16.0.20240331"
types-Pygments = "^2.18.0.20240506"
types-python-dateutil = "^2.9.0.20240821"
types-requests = "^2.32.0.20240712"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "ASYNC", "PL", "RUF"]

[tool.ruff]
line-length = 127

[tool.mypy]
disable_error_code = "import-untyped"
